+++ lib/src/StkTable/types/index.d.ts
@@ -68,7 +68,7 @@ export type StkTableColumn<T extends Record<string, any>> = {
      * - tree-node 树节点列，这一列前面有展开收起箭头
      * - selection 选择列
      */
-    type?: 'seq' | 'expand' | 'dragRow' | 'tree-node' | 'spacer' | 'selection';
+    type?: 'seq' | 'expand' | 'dragRow' | 'tree-node' | 'selection';
     /** 取值id */
     dataIndex: keyof T & string;
     /** 表头文字 */
@@ -171,17 +171,17 @@ export type PrivateStkTableColumn<T extends Record<string, any>> = StkTableColum
      * 当前列在 tableHeaderLast 中的起始叶子索引（含）。用于多级表头横向虚拟滚动。
      * @private
      */
-    __LEAF_START__?: number;
+    __LF_S__?: number;
     /**
      * 当前列在 tableHeaderLast 中的结束叶子索引（不含）。用于多级表头横向虚拟滚动。
      * @private
      */
-    __LEAF_END__?: number;
+    __LF_E__?: number;
     /**
      * 虚拟滚动右侧 spacer 占用的列数。仅用于内部 spacer 标记列。
      * @private
      */
-    __COLSPAN__?: number;
+    __VT_C_SP__?: number;
 };
 /** private row keys */
 export type PrivateRowDT = {