+++ CHANGELOG.md
@@ -9,6 +9,8 @@
     - feat: add `filter-change` emit event, triggered when filter status changes via `setFilter`.
   - feat `createEditableCell`
   - feat `createCheckboxCell`
+- optimize:
+  - Improved the start/end position algorithm for virtualX, now more precise.
   
 
 ## 0.11.15